% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/importPeakData.R
\name{importPeakData}
\alias{importPeakData}
\title{Imports data from the BVA export peaks function.}
\usage{
importPeakData(filename, pcode, conditions)
}
\arguments{
\item{filename}{A string containing the filename of the file to be imported}

\item{pcode}{A string containing the participant code}

\item{conditions}{A list of strings containing the condition labels}
}
\value{
A data frame containing properly formatted peak and area data from BVA
}
\description{
Imports data from the BVA export peaks function.
}

