% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/nemoROI.R
\name{nemoROI}
\alias{nemoROI}
\title{NEMO region of interest recoding function}
\usage{
nemoROI(elec)
}
\arguments{
\item{elec}{The 10-10 electrode label string}
}
\value{
The ROI label
}
\description{
This function will recode 10-10 electrode locations with the NEMO ROI labels
}

